requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
grpcio==1.68.1
grpcio-tools==1.68.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
protobuf==5.29.1
setuptools==75.6.0
Werkzeug==3.1.3
